requirements.txt
absl-py==2.4.0
annotated-types==0.7.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dill==0.4.1
Flask==3.1.3
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.3.0
h5py==3.14.0
idna==3.11
imbalanced-learn==0.14.1
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.12.1
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.0
rich==14.3.3
scikeras==0.13.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sklearn-compat==0.1.5
tensorflow_cpu==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.7
wrapt==2.1.2
